backend/requirements.txt
# PathShala AI — Backend Dependencies
# Python 3.11+

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Pydantic v2 + settings
pydantic==2.10.3
pydantic-settings==2.6.1

# Google Gemini AI (active LLM)
google-generativeai==0.8.3

# AWS SDK (Bedrock — preserved for future switch-back)
boto3==1.35.78
botocore==1.35.78

# HTTP client
httpx==0.27.2

# Twilio — WhatsApp delivery
twilio==9.3.7

# Environment variable loading (dev convenience)
python-dotenv==1.0.1

# AWS Lambda adapter
mangum==0.17.0
